Support CDRv3 media items
